dev -> master, Added core::Model, core::Mesh

+++ src/FileSystem.cpp
@@ -7,7 +7,7 @@ int64_t FileSystem::GetFileSize(const string& path) {
     std::ifstream file(path, std::ifstream::ate | std::ifstream::binary);
     return file.tellg();
 }
-const bgfx::Memory* FileSystem::ReadFileToMemory(const string &path, int64_t size) {
+const bgfx::Memory* FileSystem::ReadFileToMemory(const string &path, size_t size) {
     std::ifstream file(path, std::ios::binary);
 
     const bgfx::Memory* mem = bgfx::alloc(size + 1);
